refactor: split otp into OtpCodeInput+ResendCountdown; write into RequestPreview+MessageComposer
